consider sample points of nested objects

+++ src/beatmap.rs
@@ -58,8 +58,8 @@ pub struct Beatmap {
     pub circle_size: f32,
     pub overall_difficulty: f32,
     pub approach_rate: f32,
-    pub slider_multiplier: f32,
-    pub slider_tick_rate: f32,
+    pub slider_multiplier: f64,
+    pub slider_tick_rate: f64,
 
     // Events
     pub background_file: String,
